feat: 삭제시에 권한 여부 확인하고 삭제 구현

+++ src/main/java/me/siyoon/noticeboard/service/NoticeServiceImpl.java
@@ -4,13 +4,15 @@ import lombok.RequiredArgsConstructor;
 import me.siyoon.noticeboard.domain.Notice;
 import me.siyoon.noticeboard.domain.NoticeContent;
 import me.siyoon.noticeboard.domain.User;
+import me.siyoon.noticeboard.domain.enums.Authority;
 import me.siyoon.noticeboard.domain.enums.PageSize;
 import me.siyoon.noticeboard.dto.NoticeForm;
 import me.siyoon.noticeboard.repository.NoticeRepository;
 import me.siyoon.noticeboard.security.CustomUserDetails;
+import me.siyoon.noticeboard.util.UserDetailsUtil;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -37,7 +39,7 @@ public class NoticeServiceImpl implements NoticeService {
     private Notice convertFormToNotice(NoticeForm noticeForm) {
         NoticeContent noticeContent = noticeContentService.addNoticeContent(noticeForm.getContent());
 
-        CustomUserDetails userDetails = getUserDetailsFromContextHolder();
+        CustomUserDetails userDetails = UserDetailsUtil.get();
         User user = userService.getUser(userDetails.getEmail());
 
         return Notice.builder()
@@ -45,10 +47,6 @@ public class NoticeServiceImpl implements NoticeService {
                 .noticeContent(noticeContent)
                 .user(user)
                 .build();
-    }
-
-    private CustomUserDetails getUserDetailsFromContextHolder() {
-        return (CustomUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
     }
 
     @Override
@@ -66,6 +64,22 @@ public class NoticeServiceImpl implements NoticeService {
 
     @Override
     public void deleteNotice(Long id) {
-        noticeRepository.deleteById(id);
+        if (isValidAuthority(id)) {
+            noticeRepository.deleteById(id);
+            return;
+        }
+
+        // TODO 권한검사 throw InValid Autohority
+    }
+
+    private boolean isValidAuthority(Long id) {
+        Long userId = this.getNotice(id).getUser().getId();
+        CustomUserDetails userDetails = UserDetailsUtil.get();
+
+        if (userId.equals(userDetails.getId())) {
+            return true;
+        }
+
+        return userDetails.hasAuthority(Authority.ADMIN);
     }
 }
